fix: update delete user service to return a message upon success or failure

+++ users/lambda/delete-user/delete-users.services.ts
@@ -1,7 +1,21 @@
 import { userRepository } from "../shared/repository/user.repository";
 
-export const service = async ({ id }: { id: string }): Promise<void> => {
+export const service = async ({
+  id,
+}: {
+  id: string;
+}): Promise<{
+  message: string;
+}> => {
+  const verifyUser = await userRepository.findById(id);
+
+  if (!verifyUser) {
+    return { message: "User not found" };
+  }
+
   await userRepository.delete({
     id,
   });
+
+  return { message: "User deleted successfully" };
 };
